add console client
