+++ TagsCloudVisualization/TagsCloudVisualization/TagsCloudVisualization.csproj
@@ -9,6 +9,8 @@
 
     <ItemGroup>
         <PackageReference Include="Autofac" Version="6.5.0"/>
+        <PackageReference Include="CommandLineParser" Version="2.9.1"/>
+        <PackageReference Include="Stanford.NLP.POSTagger" Version="4.2.0.2"/>
         <PackageReference Include="System.Drawing.Common" Version="7.0.0-rc.2.22472.3"/>
         <PackageReference Include="WeCantSpell.Hunspell" Version="4.0.0"/>
     </ItemGroup>
