+++ TagsCloudVisualization/TagsCloudVisualization/TextInput/DefaultTextInput.cs
@@ -2,14 +2,7 @@
 
 public class DefaultTextInput : ITextInput
 {
-    private string path;
-
-    public DefaultTextInput(string path)
-    {
-        this.path = path;
-    }
-
-    public string GetInputString()
+    public string GetInputString(string path = Config.DefaultPath)
     {
         if (!File.Exists(path))
             throw new Exception("File doesn't exist");
